requirements.txt
bs4
discord
tensorflow
Pillow
